rag/requirements.txt
streamlit
fastapi
uvicorn
requests
python-dotenv
rank-bm25
faiss-cpu
sentence-transformers
groq 
rouge-score
